+++ CHANGELOG.md
@@ -15,3 +15,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Unzip files
 - Templates
 - Thymeleaf support
+- Variables Replacer